+++ tools_test/simulate_replacement/replacement_bp.py
@@ -3,6 +3,8 @@
 
 __author__ = "zhangxu"
 
+import os
+import json
 import random
 
 def raw_seq(raw_fasta):
@@ -28,24 +30,40 @@ def replacement_bp(bp, raw_fasta):
 
     ###### write the replaced bases to the new fasta ######
     new_fasta = "./data/replacement_" + str(bp) + "bp.fasta"
-    ref = ">replace_" + str(bp) + "bp.ref" + "\n"
-    pos_bp = random.randint(0, (40000 - int(bp)))
-    raw_base = info[(pos_bp):(pos_bp+bp)]
-    new_base = random_seq(bp)
+    json_file = ".".join(new_fasta.split(".")[:-1]) + ".json"
 
-    while raw_base == new_base:
+    ### whether to produce new fasta or not #########
+    if os.path.exists(json_file):
+        with open(json_file, 'r') as fjr:
+            info_record = json.load(fjr)
+    else:
+        print("++++++++++++ create a new changed fasta ++++++++++++++")
+        ref = ">replace_" + str(bp) + "bp.ref" + "\n"
+        pos_bp = random.randint(0, (40000 - int(bp)))
+        raw_base = info[(pos_bp):(pos_bp+bp)]
         new_base = random_seq(bp)
 
-    new_info = info[:pos_bp] + new_base + info[(pos_bp+bp):]
-    with open(new_fasta, 'w') as fw:
-        fw.write(ref)
-        fw.write(new_info)
+        while raw_base == new_base:
+            new_base = random_seq(bp)
 
-    ###### use the dictionary to save the variated information ######
-    info_record = {}
-    var_interval = str(pos_bp+1) + "-" + str(pos_bp + bp)
-    var_type = str(bp) + "bp"
-    info_record['info'] = [var_type, var_interval, raw_base, new_base]
-    info_record['fasta'] = new_fasta
+        new_info = info[:pos_bp] + new_base + info[(pos_bp+bp):]
+        with open(new_fasta, 'w') as fw:
+            fw.write(ref)
+            fw.write(new_info)
+
+        ##### if the length of base is more than 10 bp #####
+        ##### then it will be replaced by "-" ##############
+        if int(bp) > 10:
+            raw_base = "-"
+        ###### use the dictionary to save the variated information ######
+        info_record = {}
+        var_interval = str(pos_bp+1) + "-" + str(pos_bp + bp)
+        var_type = str(bp) + "bp"
+        info_record['info'] = [var_type, var_interval, raw_base, new_base]
+        info_record['fasta'] = new_fasta
+        
+        ###### add the record file to avoid info lost ########
+        with open(json_file, 'w') as fjw:
+            json.dump(info_record, fjw)
 
     return info_record
